+++ packages/blockchain/package.json
@@ -35,6 +35,7 @@
   "dependencies": {
     "@polkadot/api": "9.4.1",
     "@polkadot/api-contract": "9.4.1",
+    "@polkadot/util": "9.4.1",
     "crypto-browserify": "^3.12.0",
     "stream-browserify": "^3.0.0"
   },
