+++ packages/blockchain/src/index.ts
@@ -1,12 +1,2 @@
 export {Blockchain} from './Blockchain';
-export type {Bucket, BucketId, AccountId, StakingInfo} from './DDCCustomersPallet';
-export type {Amount} from './DDCStakingPallet';
-export type {ClusterId, Cluster, ClusterProps} from './DDCClustersPallet';
-export type {
-    StorageNodePublicKey,
-    CdnNodePublicKey,
-    StorageNode,
-    CdnNode,
-    StorageNodeParams,
-    CdnNodeParams,
-} from './DDCNodesPallet';
+export * from './types';
